requirements.txt
streamlit>=1.24.0
PyMuPDF>=1.25.0
plotly>=5.13.0
matplotlib>=3.7.1
fpdf>=1.7.2
requests>=2.28.2
beautifulsoup4>=4.11.2
pandas>=1.5.3
pdfplumber>=0.9.0
python-dotenv>=1.0.0 